CLI default range.To to MaxInt32
Summary:
Now CLI default range.To to MaxUint32 when task API does not specify range
upper bound. However, instance_id in db is int32, so db cannot handle
MaxUint32 correctly. Currently, handler would cap range.To to
config.InstanceCount to get around of the issue.

This diff changes the default range.To to MaxInt32, so db
can handle the case. As a result, this diff also removes
the unnecessary call to job config.

Because both server side and cli side enforces range.To to
be at most MaxInt32, this diff would not break existing
user behavior.

Reviewers: mu, apoorvaj, #peloton

Reviewed By: apoorvaj, #peloton

Subscribers: varung, amitbose, jenkins

Maniphest Tasks: T1997411

Differential Revision: https://code.uberinternal.com/D2008159