Reconcile active jobs which have out of sync mv_task_by_state entry
Summary:
in case of some jobs, the mv_task_by_state materialized view goes out of
sync with the base table. Which means that even if all tasks of that job
are terminated, they still show up in the materialized view in their previous
active state. Job runtime state is decided by querying this materialized view
and hence such jobs never get to a terminal state.

A good heuristic for batch jobs would be that if the job state has not changed
for over 24 hours, it might indicate that the MV is out of sync. So this diff
detects such stale jobs in the runtime updater code and recalculates the
task state map based on actual cached tasks for that job instead of going
to the DB. We don't do it all the time because it is potentially an expensive
operation to hold locks for ALL tasks of a job periodically on every run
 of runtime updater. So this diff does the recalculation based on the heuristic.

Such jobs will be recovered on jobmgr recovery or if job Refresh API is called
on the individual job. One idea is to have a periodic thread trying to go
through all active jobs and do it every few hours. I am on the fence about
this and might address that in the next diff if we decide to do it.

Resolves T1996201

Reviewers: #peloton, apoorvaj, zhixin

Reviewed By: zhixin

Subscribers: zhixin, jenkins

Maniphest Tasks: T1996201

Differential Revision: https://code.uberinternal.com/D2010151