Use slave agent IP for fetching logs
Summary:
To view the logs for a task, a client needs to fetch it over HTTP from the Mesos slave agent. On vcluster, the host-names that get assigned to the Mesos slaves are not DNS resolvable names. In addition, the agent's HTTP endpoint is only reachable through the mapped port on the host. Fortunately, the exact IP and port number is available in the "pid" attribute of slave agent's information.

This change adds a method to internal HostManager API to fetch the agent's information. This method is used by the BrowseSandbox method handler to determine the correct IP-address and port.

Test Plan: Unit-tests. Verified that logs for a task launched on vcluster can be fetched using Peloton CLI.

Reviewers: #peloton, mu, varung

Reviewed By: #peloton, mu, varung

Subscribers: sachins, jenkins

Maniphest Tasks: T1990095

Differential Revision: https://code.uberinternal.com/D2016893