Specify only hostname (no IP) for host maintenance
Summary: Currently, we need to specify both hostname and IP address for starting and stopping maintenance on host(s). However, since the AgentMap has been modified to hold more information (D2013301 and D2012219), we can get the IP from the AgentMap (hostmgr/host/map.go). This is addressed in the diff.

Test Plan:
1. Unit tests
2. Tested on pcluster

Reviewers: varung, #peloton

Reviewed By: varung, #peloton

Subscribers: rcharles, jenkins

Differential Revision: https://code.uberinternal.com/D2016815