Fix flaky `TestPeriodicCalculationWhenStarted`
Summary:
Changes the number of expectations to the mock from `Times(5)` to `MinTimes(5)`, since the
time perioid is set to 10ms the test can't rely on exactly how many times the
mock will be called. We don't care about about the exact number here as
long as its min of 5 times.

Reviewers: mabansal, #peloton

Reviewed By: mabansal, #peloton

Subscribers: sachins, jenkins

Maniphest Tasks: T2003273

Differential Revision: https://code.uberinternal.com/D2021009