Reconcile stale job index entries
Summary:
We have seen frequently that lucene index goes out of
sync with the actual state of the job and this results in
job.Query API returning terminal jobs as part of active
jobs query. These jobs clearly stand out because their
duration shows up in the order of 1000s of hours on the UI
and when you click on the job, all the tasks show up as
completed.

One way to fix it is to rebuild the lucene index on the
cassandra nodes periodically. I have tried this and this
seems to solve the problem most of the times but it is
better to do this in code. For BATCH jobs which are in active
state for over 24 hours, we will get job summary from
job index table and ignore the lucene index result.

Resolves T1996571, T1830099

Reviewers: #peloton, apoorvaj, zhixin

Reviewed By: #peloton, apoorvaj

Subscribers: jenkins

Maniphest Tasks: T1996571, T1830099

Differential Revision: https://code.uberinternal.com/D2040285