Fix bug in reading job_index table
Summary:
There is a bug in getJobSummaryFromIndex where we don't read the job_id
from job_index and directly call getJobSummaryFromResultMap on the query
result map. This results in the "invalid job_id" error. This error was
seen when the build was loaded on preprod and this function got called
because of a stale job entry.

The unit test did not catch it because it was depending on lucene index
not being refreshed immediately after updating the job_index base table.
So the test assumed that the lucene index will require an explicit refresh
after we set job_index entry to SUCCEEDED state, and this will simulate
stale jobs.

I have fixed the unit test to explicitly test reconcileStaleBatchJobsFromJobSummaryList
function and verified that it gets called using code coverage tool.

Resolves T2042695

Reviewers: #peloton, varung

Reviewed By: #peloton, varung

Subscribers: jenkins

Maniphest Tasks: T2042695

Differential Revision: https://code.uberinternal.com/D2056783