Add job level start and stop support
Summary:
Start/stop workflow is the same as restart. The basic logic is to copy the
config first, and treat it the same as an update.
The difference in logic is captureed in workflowStrategy

The diff also removes the taskRuntime from interface of
workflowStrategy.GetRuntimeDiff as it turns out to be not
needed.

Solves: T2031017, T2031019

Test Plan: tested on plucster

Reviewers: apoorvaj, mu, #peloton

Reviewed By: apoorvaj, #peloton

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2051707