Add `HostOfferID` to offers from host manager
Summary:
This diff add s a unique ID(v4 uuid) to every host offer from the host manager to be used by the
placement engine. This ID would need to be provided when launching tasks on the host manager.
This is required to remove potential race conditions where an offer from on old placement can
potentially be used by some other placement.  This is averted by the use of a unique ID for each offer.

This diff only adds the `HostOfferID` when offers are acquired from the host manager. In subsequent diffs
it will be added to the launch task calls as well.

Reviewers: varung, mabansal, mu, O521 Project Peloton: Add reviewers

Reviewed By: varung, O521 Project Peloton: Add reviewers

Subscribers: kejlberg, varung, jenkins

Maniphest Tasks: T2025443

Differential Revision: https://code.uberinternal.com/D2069673