Adding Defragmentation bin packing suppport for host manager
Summary:
Currently there is a big problem in production clusters where peloton cluster
gets fragmented, We have host reservation in Peloton but pre cusrsor to host reservation
we should have better bin packing by that we would not need host reservation for most
of the cases, and hosts will be free to schedule workloads and we will have increased
through put.

This diff is adding basic defragmentation bin packing to Peloton. It is also
adding bin packing interface and first fit / defrag implementations to start with.
With this we can add more implementations of bin packing algos in the future.
This diff is also adding generic implementation for multi key sorter by which we can
define the order of sorting and the result we will get is the oredred by form in the
order specified.

Test Plan:
tests added

Performance Tests
==============
+ tests/performance/perf_compare.py -f1 PERF_0.7.5_BASE -f2 PERF_0.7.5-19-gb38606fa_CURRENT
|    |   Cores |   TaskNum |   Sleep(s) | UseInsConf   | Version_x   |   Start(s)_x |   Exec(s)_x | Version_y          |   Start(s)_y |   Exec(s)_y |   Perf Change |
|  0 |    1000 |     10000 |         10 | True         | 0.7.5       |      7.24528 |     49.8063 | 0.7.5-19-gb38606fa |      6.74659 |     48.9287 |       -0.0176 |
|  1 |    1000 |     10000 |         10 | False        | 0.7.5       |      2.51094 |     43.3994 | 0.7.5-19-gb38606fa |      2.59815 |     47.404  |        0.0923 |
|  2 |    1000 |     10000 |         60 | True         | 0.7.5       |      7.10101 |    132.893  | 0.7.5-19-gb38606fa |      7.02955 |    133.756  |        0.0065 |
|  3 |    1000 |     10000 |         60 | False        | 0.7.5       |      2.58672 |    129.639  | 0.7.5-19-gb38606fa |      2.46407 |    132.958  |        0.0256 |
|  4 |    1000 |     50000 |         10 | True         | 0.7.5       |     29.1026  |    144.832  | 0.7.5-19-gb38606fa |     28.6228  |    142.729  |       -0.0145 |
|  5 |    1000 |     50000 |         10 | False        | 0.7.5       |      5.79216 |    118.93   | 0.7.5-19-gb38606fa |      6.25248 |    123.12   |        0.0352 |
|  6 |    1000 |     50000 |         60 | True         | 0.7.5       |     29.8835  |    435.539  | 0.7.5-19-gb38606fa |     28.7708  |    426.929  |       -0.0198 |
|  7 |    1000 |     50000 |         60 | False        | 0.7.5       |      5.33981 |    769.866  | 0.7.5-19-gb38606fa |      6.03177 |    403.276  |       -0.4762 |

Reviewers: mu, varung, avyas, #peloton, O521 Project Peloton: Add reviewers

Reviewed By: varung, #peloton

Subscribers: jenkins, zhixin

Differential Revision: https://code.uberinternal.com/D2050949