Improve storage unit test coverage
Summary:
A lot of datastore errors in storage could not be simulated
because we spin up a cassandra container and run unit tests.
Mocked datastore so that we could simulate error conditions
for storage.

Removed some dead code, spurious logs and added tests for
GetTaskIDsForJobAndState and GetTaskRuntimesForJobByRange

Bumped up store.go coverage to 80.6%

Reviewers: O521 Project Peloton: Add reviewers, mabansal

Reviewed By: O521 Project Peloton: Add reviewers, mabansal

Subscribers: xiaojian, jenkins

Differential Revision: https://code.uberinternal.com/D2081847