Removing empty test file from YARPC package
Summary:
Removing empty test file from YARPC package. We should anyways remove this code
in longer run and just import YARPC as a library. This test file is empty so
removing it for now. We can decide if we want to keep this package then we should write
tests cases for this package

Test Plan: N/A

Reviewers: adityacb, apoorvaj, O521 Project Peloton: Add reviewers

Reviewed By: adityacb, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2079099