Add CompareAndSetRuntime to support job write after read
Summary:
CompareAndSetRuntime would be used to support general write after read cache update.

In the current job manage code, `cachedJob.GetRuntime()` returns cached runtime directly without doing copy,
As a result, the function could fail, because when cache updates version internally, it would be propagated
to user as well. So user could not use the runtime version to do concurrency control.

An easy fix is to do a deep copy for each call of `cachedJob.GetRuntime()`, but it may incur too much memory
usage and perf degrade especially because we have several maps in cached job which grows proportionally to
instance number.

The way to solve it is to let `cachedJob.GetRuntime()` returns a read-only interface. User is forced to use
`util.ConvertJobRuntime()` if the user tries to convert the read-only interface to read-write `*job.RuntimeInfo`.
This func call does a deep copy, so changes to result of `util.ConvertJobRuntime()` would not contaminate cached
object.

Reviewers: apoorvaj, mu, O521 Project Peloton: Add reviewers

Reviewed By: apoorvaj, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T1918937

Differential Revision: https://code.uberinternal.com/D2078027