Fix test setup for a local linux dev laptop.
Summary: Peloton could not build on a local linux laptop because the timestamp protobuf type, which we use in our API, was not included onto the proto path when running the protoc command. I have change the bootstrap script to copy over the default protobuf types and changed the generate-protobuf.py script to add them to the proto path when running the protoc command.

Reviewers: #peloton, O521 Project Peloton: Add reviewers, avyas

Reviewed By: #peloton, O521 Project Peloton: Add reviewers, avyas

Subscribers: avyas, jenkins

Differential Revision: https://code.uberinternal.com/D2089283