Implement restart overwrite strategy
Summary:
Utilizing the write-after-read job cache API to implement restart overwrite.
An overwrite is only valid if the existing workflow is UPDATE and the new
one is also UPDATE.

The diff also adds a new `CompareAndSetConfig` API for cached job, which is used for update creation.
The issue with old method is, JobMgr first write config into cache & DB, then update runtime to point to the
new config. Between the config and runtime update, cached job would return the new config which is not
used by the job. It is especially issue if something fails between the config and runtime update (for example,
validation of overwriting workflow fails), because the cached config would never be used.

The issue is the config in cache is not the config in use. To fix the issue,
1. `cachedJob.GetConfig` now checks the runtime configuration version and the version of config it caches. If it is inconsistent, cache would fetch from DB to refill the cache with right version.
2. Add `CompareAndSetConfig` which directly return the updated config with new version. `Workflow` object can directly get the new version from the return result of `CompareAndSetConfig`, so it does not need to fetch the new config from cache/DB anymore.

This diff is based on D2078027

Reviewers: O521 Project Peloton: Add reviewers, apoorvaj, mu

Reviewed By: O521 Project Peloton: Add reviewers, apoorvaj

Subscribers: jenkins

Maniphest Tasks: T2031015, T2066301

Differential Revision: https://code.uberinternal.com/D2081571