JobMgr recovers jobs by job type
Summary:
JobMgr accepts a ENV var to decide which type of jobs it would run.
JobMgr uses this var to recover jobs. Now the assumption is that
JobMgr would only run one type of job.
This is a temporary workaround and should be deprecated after
storage refactor, after which recovery would not depend on the env
Var.

This diff also updates plcuster and deployment script.

Reviewers: apoorvaj, adityacb, O521 Project Peloton: Add reviewers

Reviewed By: apoorvaj, O521 Project Peloton: Add reviewers

Subscribers: avyas, jenkins

Maniphest Tasks: T1764947

Differential Revision: https://code.uberinternal.com/D2095401