Add `HostOfferID` to placements
Summary:
This diff adds the `HostOfferID` which the placement engine recieves
from the hostmanager to the placements which the placement engine sends to
the resource manager.

This ID will be used to launch the tasks. The list of `mesos.v1.OfferID offerIds` is
deprecated form the `Placement` message protobuf.

Reviewers: O521 Project Peloton: Add reviewers, varung

Reviewed By: O521 Project Peloton: Add reviewers, varung

Subscribers: jenkins

Maniphest Tasks: T2096971

Differential Revision: https://code.uberinternal.com/D2117803