Mark update failed if more instances failed during update than configured
Summary:
Mark an update to be failed if instances failed during update is more than configured.
`MaxInstanceAttempts` is used to indicate how many times an instance can fail before
it is marked as failed in an update. 0 means infinite attempts are allowed.

`MaxFailureInstances` is used to indicate how many instances in an update can fail
before the update is marked as failed. 0 means no limit.

Reviewers: O521 Project Peloton: Add reviewers, apoorvaj, chunyang.shen

Reviewed By: O521 Project Peloton: Add reviewers, apoorvaj

Subscribers: jenkins

Maniphest Tasks: T1878861

Differential Revision: https://code.uberinternal.com/D2105613