Remove client-added system labels at the time of job creation
Summary:
D2083665 adds system labels to Peloton tasks as a part of task metadata which ensures resource aggregation at resource pool and cluster levels. It adds a new behavior of reserving all `peloton.*` labels as system labels. This might result in job create failures for clients such as HRM Bridge, Michelangelo etc., which add these labels at the client end.

To ensure smooth transition, we will be scrubbing all `peloton.*` labels (from the task config) at the time of task launch. This is temporary fix on Peloton side and will be reverted to the original behavior (of failing on job create if the task config contains reserved labels) in the next release.

Reviewers: rcharles, O521 Project Peloton: Add reviewers

Reviewed By: rcharles, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T2113383

Differential Revision: https://code.uberinternal.com/D2136091