Use `HostOfferID` when launching batch and stateless tasks
Summary:
Every offer from the host manager now includes a `HostOfferID`. This change updates
the launch code path to use that id when launching tasks from the job manager to the
host manager. This should match for the launch tasks to succeed.

Updates the `LaunchTask` and `OfferOperations` hostmanager API  to take in the `HostOfferID`

Reviewers: O521 Project Peloton: Add reviewers, varung, mabansal

Reviewed By: O521 Project Peloton: Add reviewers, varung

Subscribers: sachins, zhixin, jenkins

Maniphest Tasks: T2096971

Differential Revision: https://code.uberinternal.com/D2144551