Clean-up resource-manager tasks stuck in LAUNCHING
Summary:
This change fixes a race condition between job-manager and resource-manager that leads to tasks being
stuck in resource-manager (and have resource allocated to them).
Say a job is created and for some reason resource-manager is slow to admit and place the tasks for the job.
Meanwhile if a kill is issued for job or a subset of the tasks in it, then resource-manager ignores the kills
because it isn't aware of the tasks yet. Later on it processes the tasks and asks job-manager to launch them.
But the tasks are already in KILLED state in job-manager, so nothing is done. The tasks continue to remain
in LAUNCHING state in resource-manager while holding onto allocated resources. The fix here issues
another kill task request for such tasks.

Test Plan: Added a delay (15s) in resource-manager enqueue-gangs. Created a job and then immediately killed it, or killed some tasks in it. In both cases, verified that there are no tasks stuck in LAUNCHING state in resource-manager active tasks.

Reviewers: O521 Project Peloton: Add reviewers, apoorvaj, zhixin

Reviewed By: O521 Project Peloton: Add reviewers, apoorvaj

Subscribers: jenkins

Tags: #peloton

Maniphest Tasks: T2017253

Differential Revision: https://code.uberinternal.com/D2153289