Not finding the entity in the map after dequeue is not an error
Summary:
A task is untracked when its corresponding job is untracked.
A job is untracked based on the runtime updater status update via job goal state.
So, a job may get untracked while the task has not been dequeued which will cause this error to be dumped whenever the task is dequeued.
It happens pretty regularly in production, so converting from error to debug.

Resolves T2061905

Reviewers: zhixin, O521 Project Peloton: Add reviewers

Reviewed By: zhixin, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T2061905

Differential Revision: https://code.uberinternal.com/D2162693