Add filebeat_topic to archiver logs
Summary:
Spoke with filebeat team and they said that they cannot read using
"completed_job" tag in the log and need us to specify the kafka
topic in the logline that we want to stream.

Adding filebeat_topic to the completed jobs logline as part of this
commit. We will get the filebeat topic from archiver configuration
in peloton-cluster-config.

Reviewers: O521 Project Peloton: Add reviewers, varung

Reviewed By: O521 Project Peloton: Add reviewers, varung

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2172853