Maintain all host states in Peloton
Summary:
Before, every QueryHosts call used to generate a Mesos call, to get hosts in DRAINING and DOWN states. With this change, QueryHosts calls  will be satisfied by Peloton without an explicit call to Mesos Master. We maintain a map of hostname to host state in Peloton and reconcile the host states periodically with Mesos.

This map only tracks maintenance host states (DRAINING and DOWN) and is kept separate from AgentMap (which is 'loaded' every 15s) in order to reduce the number of GetMaintenanceStatus() call to Mesos Master

Test Plan: Tested on pcluster

Reviewers: varung, apoorvaj, O521 Project Peloton: Add reviewers

Reviewed By: varung, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T2057209

Differential Revision: https://code.uberinternal.com/D2113835