Move Requeue of un-placed tasks inside rmtask
Summary:
This diff moves the business logic of dealing with an unplaced task outside
of the resource manager handler and into the RMTask where all operations
on the task can be done with the lock on the task held. This is being done to
make sure that that the mutations on an RMTask are transactional so that it
doesn't lead to any undefined behaviour. Now, once the lock is held we can
be sure that the task won't transition to any other state while it is being processed.

Reviewers: O521 Project Peloton: Add reviewers, mabansal, rcharles, O2101 Project Peloton: Resource Manager!

Reviewed By: O521 Project Peloton: Add reviewers, mabansal

Subscribers: jenkins

Maniphest Tasks: T2159921

Differential Revision: https://code.uberinternal.com/D2176455