Implement pause and resume for an update
Summary:
When an update is paused, it would return to its original state when it resumes.
For example, if a rolling forward update is paused, it would become rolling forward again after resumed.
It is same for initialized and rolling backward update.

The diff add Pause() and Resume() methods to update. Another to implement the functionality is to
use WriteProgressUpdate(). But this could have some race condition. For example, if we resume an update by:
```
prevState := update.GetStateBeforePause()
update.WriteProgress(prevState)
```
It is possible that the prevState changes between the call, so we need an atomic way to get the prevState and
persist to DB.

Reviewers: O521 Project Peloton: Add reviewers, apoorvaj

Reviewed By: O521 Project Peloton: Add reviewers, apoorvaj

Subscribers: jenkins

Maniphest Tasks: T1774741

Revert Plan: tested on pcluster

Differential Revision: https://code.uberinternal.com/D2191115