Fix update stuck when instances failed
Summary:
Previously, the code reenque update if the number of
instances done is equal to the number of instances
to be updated. However, with failed instances
introduced, these instances need to be considered
as well, or the update can get stuck.

Test Plan: tested via unit test and pcluster

Reviewers: apoorvaj, O521 Project Peloton: Add reviewers

Reviewed By: apoorvaj, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T2169101

Differential Revision: https://code.uberinternal.com/D2230847