Fix deadlock in resmgr/respool/respool.go
Summary:
As explained in the linked phab task, the deadlock occurs in the `respool.ToResourcePoolInfo()` when there is a call to `respool.GetPath()`. This is due to recursive read locking.
```
If a goroutine holds a RWMutex for reading and another goroutine might call Lock, no goroutine should expect to be able to acquire a read lock until the initial read lock is released. In particular, this prohibits recursive read locking. This is to ensure that the lock eventually becomes available; a blocked Lock call excludes new readers from acquiring the lock.
```

The sequence of calls that cause deadlock is
  1. Read lock at ToResourcePoolInfo
  2. Write Lock by SetEntitlementResources
  3. Read Lock for GetPath, it cannot get lock since it will starve write lock <deadlock>

Reviewers: varung, avyas, O521 Project Peloton: Add reviewers, amitbose

Reviewed By: varung, O521 Project Peloton: Add reviewers, amitbose

Subscribers: mabansal, jenkins, amitbose

Maniphest Tasks: T2195153

Differential Revision: https://code.uberinternal.com/D2239571