Part I: Use active_jobs table instead of mv_jobs_by_state for recovery
Summary:
Currently we do job recovery on jobmgr/resmgr restart by reading active jobs
from mv_jobs_by_state materialized view. This works for the most part but
we have a couple of problems with this approach:
1. Materialized view updates are passive and under the hood and might not
   immediately reflect the state of jobs. It is possible that some jobs
   that were just created before jobmgr restart might not be recovered. This
   has not happened so far (that we know of) but is a bigger risk for stateless jobs.
2. Sometimes MV can go out of sync with the base table. This means that some jobs
   which are terminal are still listed as active in the MV and jobmgr will
   try to recover these on every restart.
3. For stateless jobs, we need to recover all jobs. Recovering jobs for all states
   in the materialized view is not possible because query on terminal states can
   result in literally millions of batch job entries.

To solve these problems, this diff introduces active_jobs table which is
basically a synthetically sharded list of jobIDs which are active in that
cluster. This includes all batch jobs which are in one of the INITIALIZED, PENDING
or RUNNING state and ALL stateless jobs (because we recover all stateless jobs).

We will add an active job ID to this list from two places:
1. From `cached/job.go` at the time the job is created
2. From `job_runtime_updater.go` if a BATCH job moves from a terminal state to
   active state (one of the tasks of that job gets restarted)

We will remove the active job ID from this list when a BATCH job reaches
a terminal state. This will be done as a goalstate action for this jobID and
will be retried if the delete fails.

The second part of this change will cover migrating the existing active jobs
to this list, and actually using this table at the time of job recovery.

Test Plan: Added unit tests.

Reviewers: O521 Project Peloton: Add reviewers, apoorvaj, zhixin, sishi

Reviewed By: O521 Project Peloton: Add reviewers, apoorvaj

Subscribers: varung, jenkins

Differential Revision: https://code.uberinternal.com/D2289875