Delay evaluation of job by goal state if job is partially created
Summary:
Issue: If multiple job updates are requested back-to-back without
ensuring that the previous update has been successfully ececuted by
goal state engine, it can lead to all threads of the goal state engine
locked on the same job.
Correct Solution. Better handling in the goalstate engine of the evaluation
of the same entity.
Temporary solution. Delay the evaluation of the job if it is already in
INITIALIZED state. This will ensure that multiple update requests
result in only one dequeue instead of multiple dequeues.

Reviewers: zhixin, amitbose, O521 Project Peloton: Add reviewers

Reviewed By: zhixin, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2302251