Use deadline queue in the goal state engine as the queue serving the worker pool
Summary:
In the current implementation, there are two queues in the goal state engine.
The first queue is a deadline queue, and the second queue is the queue serving the worker pool.
The deadline queue has dedup functionality which ensures that each entity is present only once in the queue.
Thus, if the same entity is enqueued again, then deadline queue chooses the lowest deadline and puts it there.
However, the queue serving the worker pool has no similar dedup functionality.
Thus, if the same entity is enqueued multiple times with time.Now(), then the deadline
queue cannot dedup it, and the worker pool queue gets filled with multiple entries of the same entity.
This change removes the second queue serving the worker pool, and instead now the deadline queue
serves as the worker pool queue. Thus the dedup functionality in the deadline queue gets used
when the worker pool is completely occupied. This will reduce the chances of the same entity
occupying all the worker pool threads in a busy system.

Resolves T2287289

Reviewers: min, O521 Project Peloton: Add reviewers, zhixin

Reviewed By: min, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T2287289

Differential Revision: https://code.uberinternal.com/D2313379