Refactor JobMgr cache to guard workflow cache by job cache lock
Summary:
v1 alpha api requires concurrency control for operations like pause workflow.
Such operations need mutation on both job and workflow cache.
To make sure concurrency control is correct, the two objects need
to be guarded by the same lock.
To achieve this, all the operations on workflow that need both workflow
and job are refactored to be part of job api and are guarded by job lock.

Now workflow is part of job cache in a similar way of task.
workflow is stored as a map in job. Only one of them is active
update, and it is determined by the updateID in job runtime.
The other workflows in map can be workflows that would be aborted.
workflow map would be cleaned up in goal state engine
when the workflow is terminated.

Potential issues:
`GetUpdateProgress` was using cached job to get task runtime.
With the refactor, workflow cache should not call job cache anymore,
so the new code uses db to read task runtime directly.
The impact on perf is not clear now, maybe follow up with
some optimization if there is a need.

Follow up diffs:
1. fix job runtime change notification
2. add entity version check and concurrency control

Reviewers: O521 Project Peloton: Add reviewers, apoorvaj, sachins

Reviewed By: O521 Project Peloton: Add reviewers, apoorvaj

Subscribers: jenkins

Maniphest Tasks: T2240113

Differential Revision: https://code.uberinternal.com/D2289989