Enhance optional arguments for customized pcluster usage
Summary:
This allows us to have a local development and test environment for the
integration of Peloton and Mesos agents running in Linux VMs to facilitate
network namespaces. To do this, we need the ability to launch a pcluster
to point to a pre-existing Mesos and ZK ensemble running in local VMs.

Test Plan:
1. Bring up a local Mesos cluster using the instructions here:
   https://code.uberinternal.com/diffusion/ENMESGM/
2. Bring up a pcluster using the following arguments:
   ./tools/pcluster/pcluster.py setup --no-mesos --zk_url="192.168.33.6:2181" -a
3. Verify the Peloton framework registration
   http://192.168.33.6:5050/#/frameworks
4. Verified that the legacy pcluster behavior did not break

Reviewers: amitbose, pourchet, apoorvaj, varung, O521 Project Peloton: Add reviewers

Reviewed By: amitbose, O521 Project Peloton: Add reviewers

Subscribers: jenkins, kiranb

Differential Revision: https://code.uberinternal.com/D2315899