Add two tables (job.Get and job.Update) to Peloton performance report.
Summary:
The purpose and goal of this revision is to pick up the conversation https://code.uberinternal.com/D2255569#inline-16404413 and continue the implementation. We left the discussion on if it’s better to create multiple csv (if so, how?) or to combine all results into one csv.

I did a search on csv definition and its usage, and here is what I found: CSV is a simple file format used to store tabular data, such as a spreadsheet or database. This explanation suggests that one csv ought to serve one purpose (one set of data), and thus with three perf tests conducted in our case, each with its own purpose, we would want to have multiple csv files to store perf results.

Things that can be optimized: If more graphs are to be added to HTML Template, we can automate tables using `PERF_TEST_CONDUCTED` from `multi_benchmark.py’.

Reviewers: amitbose, sishi, adityacb, O521 Project Peloton: Add reviewers

Reviewed By: amitbose, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T2274899

Differential Revision: https://code.uberinternal.com/D2323105