Fix archiver flag, and context timeout issue
Summary:
There is an issue with the way kingpin reads booleans.
The default value for `stream` variable is set to true
in archiver's base.yaml config.
So on startup, archiver config reads this value and sets
streamOnlyMode flag to True. Now, if we set the
STREAM_ONLY_MODE environment variable to false, that is
never parsed by kingpin, and there is no way to reset this
flag that was set in the base.yaml.
The only solution is to remove this entry in base.yaml and
toggle it reliably by setting or unsetting STREAM_ONLY_MODE
env variable via per cluster config files.

Archiver has been running on preprod for a few days and
deleting jobs reliably. After this, we are ready to roll out on
sjc1-prod02 and dca1-prod02 this week.

Reviewers: O521 Project Peloton: Add reviewers, avyas, sachins

Reviewed By: O521 Project Peloton: Add reviewers, avyas

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2372119