Get task configs from task_config_v2 first
Summary:
We have enabled dual writes of task config to task_config
as well as the new task_config_v2 tables.

As part of this diff, I am forcing the GetTaskConfig and
GetTaskConfigs code to first read task_config_v2 table and
if data is not found there, read the legacy task_config
table.

I added a metric `TaskGetConfigLegacy` which is a counter
incremented every time a task config is read from the
legacy task_config table.

As soon as this metric consistently drops to 0, we will be
ready to stop dual writes to task_config and task_config_v2
and completely switch over to the new task_config_v2 table

Resolves T1935079

Reviewers: O521 Project Peloton: Add reviewers, apoorvaj, amitbose, sishi

Reviewed By: O521 Project Peloton: Add reviewers, sishi

Subscribers: jenkins

Maniphest Tasks: T1935079

Differential Revision: https://code.uberinternal.com/D2382509