Use `PREEMPTING` as goal state in job manager for batch
Summary:
Spark driver has an assumption on the task goalstate being set to `PREEMPTING`
when as task is preempted. This was changed in an earlier diff to use the same logic
as restart in the goalstate engine.  This diff reverts that change to use `PREEMPTING`
again iff
1. The task is of a job of the BATCH
2. The preemption policy is set to kill on preempt.

This is a short term fix which should satisfy the spark use case.

Test Plan:
1. Unit test
2. Added integration test

Reviewers: O521 Project Peloton: Add reviewers

Differential Revision: https://code.uberinternal.com/D2376507