Implement ORM update functionality
Summary:
Implement Update function for cassandra connector.
Added update template to generate update CQL query.
Added code to test updates with secret object

The key aspect here is the functionality to be able
to update selected few fields of a storage object in
the DB. That differentiates Update from Create.
Added a `fieldsToUpdate` list of fields to the `Update`
client interface method, so the caller can exactly
specify which fields of that object are to be updated
in the DB.

Resolves T2287239

Test Plan:
Added unit tests for client, storage object and connector
to cover Update functionality

Reviewers: O521 Project Peloton: Add reviewers, sishi, amitbose

Reviewed By: O521 Project Peloton: Add reviewers, sishi, amitbose

Subscribers: jenkins

Maniphest Tasks: T2287239

Differential Revision: https://code.uberinternal.com/D2386723