Implement PauseJobUpdate
Summary:
Simple implementation of PauseJobUpdate, which cleanly maps to PauseJobWorkflow. Not included is
marking this as a true paused state (as opposed to an awaiting pulse state) in the opaque data. I
will address opaque data / state management once we need to start differentiating between these states.

This diff also moves the unused handler methods to another file to free up some space in handler.go

Reviewers: varung, kevinxu, apoorvaj, O521 Project Peloton: Add reviewers

Reviewed By: varung, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2393007