Change git repo name to github.com/uber/peloton
Summary:
This change replaces the git repo name from
`code.uber.internal/infra/peloton` to `github.com/uber/peloton`
so that we can be ready to mirror the repo on github.com
and periodically sync commits to the public github repo.

With this change, we now need to checkout the Peloton repo
to a different local directory at
$GOPATH/src/github.com/uber/peloton. Please see the updated
README file for more details.

Test Plan: (1) unit tests and integration tests

Reviewers: #peloton, O521 Project Peloton: Add reviewers, O2101 Project Peloton: Resource Manager, amitbose, mabansal

Reviewed By: #peloton, O521 Project Peloton: Add reviewers, O2101 Project Peloton: Resource Manager, amitbose, mabansal

Subscribers: amitbose, kevinxu, jenkins

Differential Revision: https://code.uberinternal.com/D2381431